Refactor text component with improved config handling and value management
