+++ apps/application/components/Config/Renderer/text/index.tsx
@@ -33,19 +33,24 @@ export const TextInput: React.FC<Props> = ({
       {isEmpty && (
         <Text_Base value={value} onChange={(val) => setValue(val)} area />
       )}
-      {!isEmpty && (
+      {!isEmpty && value.text && (
         <Text_Base
-          value={value.value}
-          onChange={(val) => setValue((prev: any) => ({ ...prev, value: val }))}
+          value={value.text}
+          onChange={(val) =>
+            setValue((prev: any) => ({
+              ...prev,
+              text: { ...prev.text, config: val },
+            }))
+          }
         />
       )}
-      {addOn.color && (
+      {addOn.color && value.color && (
         <Color_Base
           value={value.color}
           onChange={(val) => setValue((prev: any) => ({ ...prev, color: val }))}
         />
       )}
-      {addOn.pixel && (
+      {addOn.pixel && value.pixChoice && (
         <Select_Base
           list={pixChoice}
           onChange={(val) =>
